+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.26
+
+* added `dont_raise` to `Condition`